+++ NEWS.md
@@ -1,6 +1,10 @@
 
 # dev
 
+* `print.rcmdcheck()` now has a `test_output` argument and
+  `rcmdcheck.test_output` global option, to control whether to print the full
+  test output or not. (#121)
+
 * RStudio's Pandoc is now on the path during `rcmdcheck()` 
   and `rcmdcheck_process` (#109, #132, @dpprdan).